Player View Model will represent what the player looks like to the presenter. Also, since it extends the JComponent rectangle, it will also be the physical image/hitbox of the player.
**IMPORTANT** updatePosition method should be continuously called as long as the main game is running, somewhere in the game loop should contain a call stack resulting in this method call.
