--- src/main/java/Entities/PlayerMover.java
@@ -1,5 +1,0 @@
-package Entities;
-
-public class PlayerMover {
-
-}
