+++ src/main/java/presenters/Settings.java
@@ -1,4 +1,4 @@
-package Entities;
+package presenters;
 
 public class Settings {
     /**
